This PR optimizes the library for better use in concurrent environments;
1. Mark all functions with `schedule = "DirtyCpu")` to ensure the main Beam scheduler does not wait on rather slow tiktoken functions. NIFs should generally execute in under 1 ms for the scheduler work as intended. This can lead to scheduler collapse (more on [that here](https://gist.github.com/chewbranca/07d9a6eed3da7b490b47)). Marking the functions as "DirtyCpu" elevates this a bit as a separate scheduler will be used. 
2. Instead of using a singleton tokenizer, create one per thread using thread locals. This performs marginally better than creating a new tokenizer on demand and much better than a singleton.


Benchee results for 16 parallel execution on 8 cores to tokenize 1 million words:
```
Name                          ips        average  deviation         median         99th %
count thread local           1.33      752.04 ms    ±12.26%      735.96 ms     1049.66 ms
encode thread local          1.30      771.91 ms     ±6.77%      769.04 ms      885.09 ms
encode no cache              1.17      853.15 ms     ±6.05%      851.28 ms      972.03 ms
encode singleton             0.22     4452.14 ms    ±33.68%     5314.44 ms     5374.20 ms

Comparison: 
count thread local           1.33
encode thread local          1.30 - 1.03x slower +19.88 ms
encode no cache              1.17 - 1.13x slower +101.11 ms
encode singleton             0.22 - 5.92x slower +3700.11 ms
```